+++ src/django/api/management/commands/makefixtures.py
@@ -288,6 +288,7 @@ def make_facility_list_item(source_pk, item_pk, row_index, raw_data):
             'raw_data': raw_data,
             'status': 'UPLOADED',
             'processing_results': [],
+            'sector': [],
             'created_at': created_at,
             'updated_at': updated_at,
         }